+++ src/config/mod.rs
@@ -1,0 +1,4 @@
+mod config;
+mod defaults;
+
+pub use config::*;
